Fixes #492. Deprecate Processor.CONTENT_MODEL_* and moved @ContentModel to org.asciidoctor.ast.
