+++ asciidoctorj-documentation/src/test/java/org/asciidoctor/integrationguide/extension/YellBlockProcessor.java
@@ -2,7 +2,7 @@ package org.asciidoctor.integrationguide.extension;
 
 import org.asciidoctor.ast.StructuralNode;
 import org.asciidoctor.extension.BlockProcessor;
-import org.asciidoctor.extension.ContentModel;
+import org.asciidoctor.ast.ContentModel;
 import org.asciidoctor.extension.Contexts;
 import org.asciidoctor.extension.Name;
 import org.asciidoctor.extension.Reader;
